# shared/api-contract/openapi/v1/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/hotels": {
      "post": {
        "operationId": "HotelsController_createHotel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tenantId": {
                    "type": "string",
                    "maxLength": 80
                  },
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 160
                  },
                  "timezone": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 80,
                    "default": "Asia/Saigon"
                  },
                  "brandSettings": {
                    "type": "object",
                    "additionalProperties": true
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đã tạo khách sạn",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 201
                    },
                    "error": {
                      "type": "object",
                      "nullable": true,
                      "additionalProperties": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo khách sạn thành công"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "tenantId": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "timezone": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "brandSettings": {
                          "type": "object",
                          "additionalProperties": true,
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "tenant": {
                          "type": "object",
                          "nullable": true,
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "code": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "updatedAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          },
                          "required": [
                            "id",
                            "code",
                            "name",
                            "createdAt",
                            "updatedAt"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "tenantId",
                        "name",
                        "code",
                        "timezone",
                        "status",
                        "brandSettings",
                        "createdAt",
                        "updatedAt",
                        "tenant"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "error",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "hotels"
        ]
      },
      "get": {
        "operationId": "HotelsController_listHotels",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "description": "Tenant scope chính; tenantId query chỉ là fallback tương thích tạm thời.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Đã lấy danh sách khách sạn",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "number",
                      "example": 200
                    },
                    "error": {
                      "type": "object",
                      "nullable": true,
                      "additionalProperties": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy danh sách khách sạn thành công"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "total": {
                          "type": "number"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "tenantId": {
                                "type": "string"
                              },
                              "name": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              },
                              "timezone": {
                                "type": "string"
                              },
                              "status": {
                                "type": "string"
                              },
                              "brandSettings": {
                                "type": "object",
                                "additionalProperties": true,
                                "nullable": true
                              },
                              "createdAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "tenant": {
                                "type": "object",
                                "nullable": true,
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "code": {
                                    "type": "string"
                                  },
                                  "name": {
                                    "type": "string"
                                  },
                                  "createdAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  },
                                  "updatedAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  }
                                },
                                "required": [
                                  "id",
                                  "code",
                                  "name",
                                  "createdAt",
                                  "updatedAt"
                                ]
                              }
                            },
                            "required": [
                              "id",
                              "tenantId",
                              "name",
                              "code",
                              "timezone",
                              "status",
                              "brandSettings",
                              "createdAt",
                              "updatedAt",
                              "tenant"
                            ]
                          }
                        }
                      },
                      "required": [
                        "page",
                        "limit",
                        "total",
                        "items"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "error",
                    "message",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "hotels"
        ]
      }
    },
...
    "/hotels/{hotelId}/folios": {
      "get": {
        "operationId": "FolioController_listFolios",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "OPEN",
                "CLOSED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách folio"
          }
        },
        "tags": [
          "billing-folios"
        ]
      }
    },
    "/hotels/{hotelId}/stays/{stayId}/active-folio": {
      "get": {
        "operationId": "FolioController_getActiveFolioByStay",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stayId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Folio đang mở"
          }
        },
        "tags": [
          "billing-folios"
        ]
      }
    },
    "/hotels/{hotelId}/folios/{folioId}/summary": {
      "get": {
        "operationId": "FolioController_getFolioSummary",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "folioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tổng quan folio"
          }
        },
        "tags": [
          "billing-folios"
        ]
      }
    },
    "/hotels/{hotelId}/folios/{folioId}/checkout/issue-invoice": {
      "post": {
        "operationId": "FolioController_issueInvoice",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "folioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice đã phát hành"
          }
        },
        "tags": [
          "billing-folios"
        ]
      }
    },
    "/hotels/{hotelId}/folios/{folioId}/items": {
      "get": {
        "operationId": "FolioController_listFolioItems",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "folioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách FolioItem"
          }
        },
        "tags": [
          "billing-folios"
        ]
      }
    },
    "/hotels/{hotelId}/folios/{folioId}": {
      "get": {
        "operationId": "FolioController_getFolioDetail",
        "parameters": [
          {
            "name": "hotelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "folioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết folio"
          }
        },
        "tags": [
          "billing-folios"
        ]
      }
    },
...
  },
  "info": {
    "title": "VietSage Auth Service API",
    "description": "HTTP API contract for auth, rbac, and hotel-users modules",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}